Move ParameterCanBeByValInspectionTests for QuickFixes into relevant files under RD.QuickFixes
